Security Fix - Access Levels Read From com_events Sections
Reads The Access Levels set from admin console on who should see items
and do actions from the controller. No access can be granted to
non-access group members

+++ com_events/site/controllers/delete.php
@@ -42,23 +42,31 @@
 							{
 								return JError::raiseError(403, JText::_('COM_EVENTS_ERROR_TEAM_ACTION_FORBIDDEN'));
 							}*/
-							
 							// Gets required variables for update.
 							$team = JRequest::getInt('id');
-							if(JFactory::getUser()->guest || empty($team))
+							
+							// Does a security check to verify access is allowed to this team
+							if(in_array($model->getTeam($team)->access, JAccess::getAuthorisedViewLevels(JFactory::getUser()->id))) 
 							{
-								$app->enqueueMessage(JText::_('COM_EVENTS_ERROR_LOGIN_REQUIRED'), 'error');
-								$url = (JRoute::_('index.php?option=com_events&view=team&id=' . (int) $team . '&layout=delete&action=failure', false));							
-							}
-							else if($model->deleteTeam($team))
-							{
-								$url = (JRoute::_('index.php?option=com_events&view=team&id=' . (int) $team . '&layout=delete&action=success', false)); 
+							
+								if(JFactory::getUser()->guest || empty($team))
+								{
+									$app->enqueueMessage(JText::_('COM_EVENTS_ERROR_LOGIN_REQUIRED'), 'error');
+									$url = (JRoute::_('index.php?option=com_events&view=team&id=' . (int) $team . '&layout=delete&action=failure', false));							
+								}
+								else if($model->deleteTeam($team))
+								{
+									$url = (JRoute::_('index.php?option=com_events&view=team&id=' . (int) $team . '&layout=delete&action=success', false)); 
+								}
+								else
+								{
+									$url = (JRoute::_('index.php?option=com_events&view=team&id=' . (int) $team . '&layout=delete&action=failure', false)); 
+								}
 							}
 							else
 							{
 								$url = (JRoute::_('index.php?option=com_events&view=team&id=' . (int) $team . '&layout=delete&action=failure', false)); 
 							}
-
 							$app->redirect($url);
 							break;
 					}
